claude_spikes/requirements_experiments.txt
# Additional dependencies for AWS registration experiments
# Add these to your main pyproject.toml as needed

playwright>=1.40.0
selenium>=4.15.0
beautifulsoup4>=4.12.0
requests>=2.31.0
selenium-stealth>=1.0.6

# For comparison testing
pytest-asyncio>=0.21.0